worked on some actor ai wich is absolutely TERRIBLE! Gonna freeze the game for like 10 seconds if I try to move more than a few squares, but it might just work. Probably not though. Wait, let me try something: \n! did it work? We shall see. Also I think adding the healing stones and ring thingy was the other thing I did today.

+++ com/canfield010/mygame/item/Backpack.java
@@ -1,11 +1,11 @@
 package com.canfield010.mygame.item;
 
-import com.canfield010.mygame.mapsquares.MapSquare;
+import com.canfield010.mygame.mapsquare.MapSquare;
 
 public class Backpack extends Item {
 
     public Backpack() {
-        super("Backpack", 1);
+        super("Backpack", (byte)1);
     }
 
     @Override
