+++ com/canfield010/mygame/item/Coal.java
@@ -1,10 +1,10 @@
 package com.canfield010.mygame.item;
 
-import com.canfield010.mygame.mapsquares.MapSquare;
+import com.canfield010.mygame.mapsquare.MapSquare;
 
 public class Coal extends Item {
     public Coal() {
-        super("Coal", 32);
+        super("Coal", (byte)32);
     }
 
     @Override
