+++ com/canfield010/mygame/item/Log.java
@@ -1,10 +1,10 @@
 package com.canfield010.mygame.item;
 
-import com.canfield010.mygame.mapsquares.MapSquare;
+import com.canfield010.mygame.mapsquare.MapSquare;
 
 public class Log extends Item {
     public Log() {
-        super("Log", 32);
+        super("Log", (byte)32);
     }
 
     @Override
