+++ com/canfield010/mygame/item/RawBeef.java
@@ -1,10 +1,10 @@
 package com.canfield010.mygame.item;
 
-import com.canfield010.mygame.mapsquares.MapSquare;
+import com.canfield010.mygame.mapsquare.MapSquare;
 
 public class RawBeef extends Item {
     public RawBeef() {
-        super("Raw Beef", 32);
+        super("Raw Beef", (byte)32);
     }
 
     @Override
